// frontend/src/redux/features/types.d.ts
export type TypeItemTableSaveTemplete = {
  body_text: string;
  category: string;
  company_name: string;
  footer_text: string;
  id: string;
  image_handle: string;
  image_url: string;
  language: string;
  name: string;
  phone_number_id: string;
  status: string;
  template_id: string;
};